ajustando codigo
