+++ src/components/cards/cardDefinicaoFase/cardDefinicaoFase.module.css
@@ -3,24 +3,10 @@
   color: #222222;
   background-color: #fff;
   border-radius: 10px;
-
   justify-content: center;
   align-items: center;
   flex-direction: column;
   position: relative;
-}
-
-.tagStatusProblem {
-  position: absolute;
-  top: 0;
-  right: 0;
-
-  font-size: 12px;
-  font-weight: 500;
-  padding: 8px;
-
-  border-bottom-left-radius: 10px;
-  border-top-right-radius: 10px;
 }
 
 .headerContainer {
@@ -59,31 +45,13 @@
   margin-bottom: 1rem;
 }
 
-.footerContainer {
-  height: 48px;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  border-top: 1px solid #e8e8e8;
-  border-top-left-radius: 5px;
-  border-top-right-radius: 5px;
-  padding: 15px;
-  gap: 32px;
-}
-
-.footerContainerType2 {
+.buttonContainer {
   display: flex;
   justify-content: space-between;
   height: 55px;
   padding: 0 35px;
   border-radius: 0px 0px 8px 8px;
   border: 1px solid #bdbdbd;
-}
-
-.footerContainer svg {
-  max-height: 100% !important;
 }
 
 .divisoria {
